Fixed x86_64 orkaudio compilation error.
git-svn-id: svn://svn.code.sf.net/p/oreka/svn/trunk@603 09dcff7a-b715-0410-9601-b79a96267cd0

+++ orkaudio/audiocaptureplugins/voip/PacketHeaderDefs.cpp
@@ -154,7 +154,7 @@ bool SkinnyValidateCallInfo(SkCallInfoStruct* sci, u_char* packetEnd)
 
 bool SkinnyValidateCcm5CallInfo(SkCcm5CallInfoStruct *sci, u_char* packetEnd)
 {
-	int partiesLen = 0;
+	long partiesLen = 0;
 	bool valid = true;
 	if(((u_char*)sci + sizeof(SkCcm5CallInfoStruct)) > packetEnd)
 	{
@@ -164,7 +164,7 @@ bool SkinnyValidateCcm5CallInfo(SkCcm5CallInfoStruct *sci, u_char* packetEnd)
 	{
 		valid = false;
 	}
-	partiesLen = (int)packetEnd - (int)sci - sizeof(SkCcm5CallInfoStruct);
+	partiesLen = (long)packetEnd - (long)sci - sizeof(SkCcm5CallInfoStruct);
 	if(valid)
 	{
 		valid = checkPartyString(sci->parties, partiesLen);
